requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.6
django-bootstrap-form==3.4
django-widget-tweaks==1.4.2
idna==2.7
oauthlib==2.1.0
Pillow==5.1.0
PyJWT==1.6.4
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.23
wheel==0.24.0
django-tastypie==0.14.1
